pi/requirements.txt
requests
numpy
sounddevice
Pillow
rplidar-roboticia
pyserial
RPi.GPIO
picamera2
adafruit-circuitpython-rplidar
